+++ src/libanalyze/analyze_private.h
@@ -50,8 +50,8 @@ extern void analyze_gen_worker(int cpu, void *ptr);
 ANALYZE_EXPORT extern int analyze_get_bbox_rays(fastf_t **rays, point_t min, point_t max, struct bn_tol *tol);
 ANALYZE_EXPORT extern int analyze_get_scaled_bbox_rays(fastf_t **rays, point_t min, point_t max, fastf_t ratio);
 
-ANALYZE_EXPORT extern int analyze_get_solid_partitions(struct bu_ptbl *results, struct rt_gen_worker_vars *pstate, fastf_t *rays, int ray_cnt,
-	struct db_i *dbip, const char *obj, struct bn_tol *tol, int ncpus, int filter);
+ANALYZE_EXPORT extern int analyze_get_solid_partitions(struct bu_ptbl *results, struct rt_gen_worker_vars *pstate, fastf_t *rays, size_t ray_cnt,
+	struct db_i *dbip, const char *obj, struct bn_tol *tol, size_t ncpus, int filter);
 
 typedef struct xray * (*getray_t)(void *ptr);
 typedef int *         (*getflag_t)(void *ptr);
