+++ src/libbu/vls.c
@@ -674,7 +674,7 @@ bu_vls_gets(struct bu_vls *vp, FILE *fp)
     if (endlen < startlen)
 	return -1;
 
-    return endlen;
+    return (int)endlen;
 }
 
 void
@@ -897,7 +897,8 @@ bu_vls_simplify(struct bu_vls *vp, const char *keep, const char *de_dup, const c
 	ccnt = 0;
 	c = (unsigned char *)&(bu_vls_addr(&tmpstr)[(strlen(bu_vls_addr(&tmpstr)) - 1)]);
 	while (*c && vls_char_in_set((const char *)c, trim)) {ccnt++; c--;}
-	if (ccnt) bu_vls_trunc(&tmpstr, bu_vls_strlen(&tmpstr) - ccnt);
+	if (ccnt)
+	    bu_vls_trunc(&tmpstr, (int)(bu_vls_strlen(&tmpstr) - ccnt));
     }
 
     ret = (!bu_vls_strcmp(&tmpstr, vp)) ? 0 : 1;
@@ -991,8 +992,8 @@ int
 bu_vls_incr(struct bu_vls *name, const char *regex_str, const char *incr_spec, bu_vls_uniq_t ut, void *data)
 {
     int ret = 0;
-    int i = 0;
-    int j = 0;
+    size_t i = 0;
+    size_t j = 0;
     int offset = 0;
     regmatch_t *incr_substrs;
     regmatch_t *num_substrs;
@@ -1034,9 +1035,11 @@ bu_vls_incr(struct bu_vls *name, const char *regex_str, const char *incr_spec, b
 	    regfree(&compiled_regex);
 	}
 	i = bu_vls_strlen(name);
-	while(incr_substrs[i].rm_so == -1 || incr_substrs[i].rm_eo == -1) i--;
+	while(incr_substrs[i].rm_so == -1 || incr_substrs[i].rm_eo == -1)
+	    i--;
 
-	if (i != 1) return -1;
+	if (i != 1)
+	    return -1;
 
 	/* Now we know where the incrementer is - process, find the number, and assemble the new string */
 	bu_vls_trunc(&new_name, 0);
