+++ src/libged/tire.c
@@ -1782,7 +1782,7 @@ MakeAirRegion(struct rt_wdb (*file), char *suffix, fastf_t dyhub, fastf_t zhub, 
 }
 
 HIDDEN int
-_opt_tire_iso(struct bu_vls *msg, int argc, const char **argv, void *set_var)
+_opt_tire_iso(struct bu_vls *msg, size_t argc, const char **argv, void *set_var)
 {
     int d1, d2, d3;
     char s1, s2;
