+++ src/sig/d-i.c
@@ -47,7 +47,7 @@ main(int argc, char *argv[])
     double ibuf[512];
     short obuf[512];
 
-    int i, num;
+    size_t i, num;
     double scale = 1.0;
     double value;
     int clip_high, clip_low;
